feat(kit)!: `InputSlider` remove all deprecated API
