+++ projects/demo/src/modules/components/input-slider/examples/3/index.html
@@ -1,6 +1,5 @@
 <div class="wrapper">
     <tui-input-slider
-        new
         [min]="min"
         [max]="max"
         [steps]="totalSteps"
