+++ projects/demo/src/modules/components/input-slider/examples/5/index.html
@@ -1,6 +1,5 @@
 <section class="control">
     <tui-input-slider
-        new
         tuiTextfieldSize="m"
         [min]="min"
         [max]="max"
