+++ projects/demo/src/modules/components/input-slider/input-slider.template.html
@@ -17,29 +17,6 @@
             </a>
         </p>
 
-        <tui-notification status="warning">
-            <p
-                i18n
-                class="tui-space_top-0"
-            >
-                This component is being refactored. Soon&nbsp;(next&nbsp;major&nbsp;release) you will see the fresh
-                version of it!
-            </p>
-            <p i18n>
-                Of course, we keep backward compatibility in mind (for 2.x.x). You can still use old version of
-                <code>InputSlider</code>
-                .
-            </p>
-
-            <p i18n>
-                However, if you are going to use this component, we recommend to use new version. To enable the "new
-                version"-mode, add
-                <code>new</code>
-                directive. Example:&nbsp;
-                <code>&lt;tui&#8209;input&#8209;slider&nbsp;new&nbsp;...&gt;</code>
-            </p>
-        </tui-notification>
-
         <tui-doc-example
             id="base"
             i18n-heading
@@ -130,7 +107,6 @@
         <tui-doc-demo [control]="control">
             <ng-template>
                 <tui-input-slider
-                    new
                     [formControl]="control"
                     [min]="min"
                     [max]="max"
@@ -151,16 +127,12 @@
                     [tuiTextfieldCleaner]="cleaner"
                     [tuiTextfieldCustomContent]="customContentSelected"
                     [tuiTextfieldExampleText]="exampleText"
+                    [tuiTextfieldIconLeft]="iconLeft"
+                    [tuiTextfieldLabelOutside]="labelOutside"
                     [tuiTextfieldSize]="size"
                     [tuiHintContent]="hintContent"
                     [tuiHintDirection]="hintDirection"
                     [tuiHintMode]="hintMode"
-                    [size]="size"
-                    [pluralize]="pluralize"
-                    [segmentsPluralize]="segmentsPluralize"
-                    [minLabel]="minLabel"
-                    [maxLabel]="maxLabel"
-                    [secondary]="secondary"
                 >
                     Range
                 </tui-input-slider>
@@ -271,110 +243,6 @@
                 [(documentationPropertyValue)]="postfix"
             >
                 A postfix symbol, like currency
-            </ng-template>
-            <ng-template
-                i18n
-                documentationPropertyName="minLabel"
-                documentationPropertyMode="input"
-                documentationPropertyType="string"
-                [documentationPropertyDeprecated]="true"
-                [documentationPropertyValues]="minLabelVariants"
-                [(documentationPropertyValue)]="minLabel"
-            >
-                Label for min value.
-
-                <p>
-                    Use
-                    <code>valueContent</code>
-                    instead.
-                </p>
-            </ng-template>
-            <ng-template
-                i18n
-                documentationPropertyName="maxLabel"
-                documentationPropertyMode="input"
-                documentationPropertyType="string"
-                [documentationPropertyDeprecated]="true"
-                [documentationPropertyValues]="maxLabelVariants"
-                [(documentationPropertyValue)]="maxLabel"
-            >
-                Label for max value.
-
-                <p>
-                    Use
-                    <code>valueContent</code>
-                    instead.
-                </p>
-            </ng-template>
-            <ng-template
-                i18n
-                documentationPropertyName="size"
-                documentationPropertyMode="input"
-                documentationPropertyType="TuiSizeL"
-                [documentationPropertyDeprecated]="true"
-                [documentationPropertyValues]="sizeVariants"
-                [(documentationPropertyValue)]="size"
-            >
-                Size
-                <p>
-                    Use
-                    <code>tuiTextfieldSize</code>
-                    instead.
-                </p>
-            </ng-template>
-            <ng-template
-                i18n
-                documentationPropertyName="secondary"
-                documentationPropertyMode="input"
-                documentationPropertyType="string"
-                [documentationPropertyDeprecated]="true"
-                [documentationPropertyValues]="secondaryVariants"
-                [(documentationPropertyValue)]="secondarySelected"
-            >
-                Right label.
-                <p>
-                    Use
-                    <code>tuiTextfieldCustomContent</code>
-                    instead.
-                </p>
-            </ng-template>
-            <ng-template
-                i18n
-                documentationPropertyName="pluralize"
-                documentationPropertyMode="input"
-                documentationPropertyType="TuiPluralize | Record<string, string>"
-                [documentationPropertyDeprecated]="true"
-                [documentationPropertyValues]="pluralizeVariants"
-                [(documentationPropertyValue)]="pluralize"
-            >
-                Plural forms for labels. TuiPluralize array is deprecated. Use object that mimics the
-                <a
-                    tuiLink
-                    href="https://unicode-org.github.io/icu/userguide/format_parse/messages/"
-                >
-                    ICU format
-                </a>
-                for i18nPlural.
-
-                <p>
-                    Use
-                    <code>postfix</code>
-                    instead.
-                </p>
-            </ng-template>
-            <ng-template
-                i18n
-                documentationPropertyName="segmentsPluralize"
-                documentationPropertyMode="input"
-                documentationPropertyType="TuiPluralize"
-                [documentationPropertyValues]="pluralizeVariants"
-                [documentationPropertyDeprecated]="true"
-                [(documentationPropertyValue)]="segmentsPluralize"
-            >
-                Plural forms for
-                <code>segments</code>
-
-                <p>See examples how create ticks without this property (outside of the component).</p>
             </ng-template>
         </tui-doc-documentation>
         <inherited-documentation></inherited-documentation>
