+++ projects/demo/src/modules/components/pagination/examples/3/index.html
@@ -1,11 +1,14 @@
 <tui-input-slider
-    secondary="sidePadding"
-    size="m"
+    tuiTextfieldSize="m"
     class="slider"
+    [tuiTextfieldCustomContent]="inputCustomContent"
+    [tuiTextfieldLabelOutside]="true"
     [max]="6"
     [quantum]="1"
     [(ngModel)]="sidePadding"
 ></tui-input-slider>
+
+<ng-template #inputCustomContent>sidePadding</ng-template>
 
 <tui-pagination
     [sidePadding]="sidePadding"
