+++ projects/demo/src/modules/components/pagination/pagination.module.ts
@@ -3,6 +3,7 @@ import {NgModule} from '@angular/core';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {RouterModule} from '@angular/router';
 import {TuiAddonDocModule, tuiGenerateRoutes} from '@taiga-ui/addon-doc';
+import {TuiTextfieldControllerModule} from '@taiga-ui/core';
 import {
     TuiBadgeModule,
     TuiInputSliderModule,
@@ -27,6 +28,7 @@ import {ExampleTuiPaginationComponent} from './pagination.component';
         FormsModule,
         ReactiveFormsModule,
         TuiAddonDocModule,
+        TuiTextfieldControllerModule,
         RouterModule.forChild(tuiGenerateRoutes(ExampleTuiPaginationComponent)),
         PolymorpheusModule,
     ],
