+++ projects/demo/src/modules/markup/form/example/index.style.less
@@ -17,3 +17,10 @@
     flex: 1;
     justify-content: space-between;
 }
+
+.ticks-labels {
+    .tui-slider-ticks-labels();
+
+    color: var(--tui-text-02);
+    margin-top: 0.25rem;
+}
