+++ projects/kit/components/input-slider/input-slider.module.ts
@@ -1,31 +1,23 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {TuiFocusableModule, TuiRepeatTimesModule} from '@taiga-ui/cdk';
-import {TuiFormatNumberPipeModule, TuiTextfieldControllerModule} from '@taiga-ui/core';
+import {FormsModule} from '@angular/forms';
+import {TuiFocusableModule} from '@taiga-ui/cdk';
 import {TuiInputNumberModule} from '@taiga-ui/kit/components/input-number';
 import {TuiSliderModule} from '@taiga-ui/kit/components/slider';
 import {PolymorpheusModule} from '@tinkoff/ng-polymorpheus';
 
-import {
-    TuiInputSliderComponent,
-    TuiNewInputSliderDirective,
-} from './input-slider.component';
+import {TuiInputSliderComponent} from './input-slider.component';
 
 @NgModule({
     imports: [
         CommonModule,
         FormsModule,
-        ReactiveFormsModule,
-        TuiFocusableModule,
-        TuiSliderModule,
-        TuiFormatNumberPipeModule,
-        TuiRepeatTimesModule,
         PolymorpheusModule,
+        TuiFocusableModule,
         TuiInputNumberModule,
-        TuiTextfieldControllerModule,
+        TuiSliderModule,
     ],
-    declarations: [TuiInputSliderComponent, TuiNewInputSliderDirective],
-    exports: [TuiInputSliderComponent, TuiNewInputSliderDirective],
+    declarations: [TuiInputSliderComponent],
+    exports: [TuiInputSliderComponent],
 })
 export class TuiInputSliderModule {}
