+++ projects/kit/components/input-slider/input-slider.style.less
@@ -4,31 +4,15 @@
     .createStackingContext();
     display: block;
     height: max-content;
-
-    &._segmented._show-ticks-labels {
-        // TODO: remove in v3.0
-        /* stylelint-disable-next-line */
-        border-bottom: 26px solid transparent;
-    }
 }
 
-// TODO: remove wrapper and move all styles to input[type='range'] in v3.0
-.t-slider-wrapper {
+// Dont change CSS-selector (we require the such specificity to override default styles without !important)
+input[type='range'] {
     position: absolute;
     top: 100%;
     left: calc(var(--tui-radius-m) / 2);
     right: 0;
     margin: -0.5625rem 0 0;
-}
-
-// Dont change CSS-selector (we require the such specificity to override default styles without !important)
-input[type='range'] {
     border-top-left-radius: 0;
     border-bottom-left-radius: calc(var(--tui-radius-m) * 5) calc(var(--tui-radius-m) * 1.5);
-}
-
-// TODO: remove in v3.0
-.t-ticks-labels {
-    .tui-slider-ticks-labels(m);
-    color: var(--tui-text-02);
 }
