+++ projects/kit/components/input-slider/input-slider.template.html
@@ -4,10 +4,7 @@
     [precision]="precision"
     [decimal]="decimal"
     [prefix]="prefix"
-    [postfix]="postfix || (value | i18nPlural: pluralizeMap || pluralizeMapFallback)"
-    [tuiTextfieldCustomContent]="controller.customContent || deprecatedSecondary"
-    [tuiTextfieldSize]="computedSize"
-    [tuiTextfieldLabelOutside]="computedSize !== 'l'"
+    [postfix]="postfix"
     [readOnly]="readOnly"
     [pseudoFocus]="computedFocused"
     [pseudoHover]="pseudoHover"
@@ -21,50 +18,25 @@
 >
     <ng-content></ng-content>
     <div
-        *ngIf="computedValueContent && !focused"
+        *ngIf="valueContent && !focused"
         ngProjectAs="tuiContent"
         automation-id="tui-input-slider__value-content"
     >
-        <ng-container *polymorpheusOutlet="computedValueContent as text; context: {$implicit: value}">
+        <ng-container *polymorpheusOutlet="valueContent as text; context: {$implicit: value}">
             {{ text }}
         </ng-container>
     </div>
 </tui-input-number>
 
-<!--TODO: remove in v3.0-->
-<ng-template #deprecatedSecondary>{{ secondary }}</ng-template>
-
-<!--TODO: remove wrapper + ticks labels (leave only slider) in v3.0-->
-<div class="t-slider-wrapper">
-    <input
-        tuiSlider
-        type="range"
-        [tuiFocusable]="false"
-        [max]="computedSteps"
-        [segments]="segments"
-        [keySteps]="computedKeySteps"
-        [attr.disabled]="readOnly || disabled || null"
-        [ngModel]="value"
-        (keyStepsInput)="onSliderChange($event)"
-        (click)="focusTextInput()"
-    />
-
-    <!--TODO delete ticks labels in v3.0-->
-    <div
-        *ngIf="segments && !isNew"
-        class="t-ticks-labels"
-    >
-        <span
-            *tuiRepeatTimes="let tickIndex of segments + 1"
-            automation-id="tui-slider__segment"
-        >
-            <span *ngIf="fromToTexts$ | async as fromToText">
-                {{ getTickPrefix(tickIndex, segments, fromToText) }}
-                {{ tickIndex | tuiSliderTickLabel: segments:computedKeySteps | tuiFormatNumber }}
-                <span *ngIf="segmentsPluralizeMap">
-                    {{ tickIndex | tuiSliderTickLabel: segments:computedKeySteps | i18nPlural: segmentsPluralizeMap }}
-                </span>
-            </span>
-        </span>
-    </div>
-</div>
+<input
+    tuiSlider
+    type="range"
+    [tuiFocusable]="false"
+    [max]="computedSteps"
+    [segments]="segments"
+    [keySteps]="computePureKeySteps(keySteps, min, max)"
+    [attr.disabled]="readOnly || disabled || null"
+    [ngModel]="value"
+    (keyStepsInput)="onSliderChange($event)"
+    (click)="focusTextInput()"
+/>
